# Improve performance and reduce allocations in RouteValueDictionary.